Add setCreditScore to saveCustomer and updateCustomer at CustomerService

+++ loan-application-system/src/main/java/dev/patika/loanapplicationsystem/entity/CreditScore.java
@@ -16,5 +16,5 @@ import javax.persistence.OneToOne;
 public class CreditScore extends AbstractBaseEntity{
     private int creditScore;
     //@OneToOne
-    private int lastDigitOfIdNumber;
+    private long lastDigitOfIdNumber;
 }
